# api-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for reading device telemetry and aggregate metrics.",
        "title": "AgroIoT API Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/devices/{device_id}/aggregates": {
            "get": {
                "description": "Returns avg, min, and max values for a device metric in the requested time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telemetry"
                ],
                "summary": "Get metric aggregates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "temperature",
                        "description": "Metric name",
                        "name": "metric",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2026-05-10T00:00:00Z",
                        "description": "Start time in RFC3339 format",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-05-10T23:59:59Z",
                        "description": "End time in RFC3339 format",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/devices/{device_id}/aggregates/timeseries": {
            "get": {
                "description": "Returns avg, min, and max values grouped by hour or day for a device metric.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telemetry"
                ],
                "summary": "Get time-series metric aggregates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "temperature",
                        "description": "Metric name",
                        "name": "metric",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "hour",
                            "day"
                        ],
                        "type": "string",
                        "default": "hour",
                        "description": "Aggregation interval",
                        "name": "interval",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-05-10T00:00:00Z",
                        "description": "Start time in RFC3339 format",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-05-10T23:59:59Z",
                        "description": "End time in RFC3339 format",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/devices/{device_id}/telemetry": {
            "get": {
                "description": "Returns telemetry records for a device in the requested time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telemetry"
                ],
                "summary": "Get device telemetry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2026-05-10T00:00:00Z",
                        "description": "Start time in RFC3339 format",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-05-10T23:59:59Z",
                        "description": "End time in RFC3339 format",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}